working on deals module api integration
