+++ package.json
@@ -28,6 +28,7 @@
     "react-router-dom": "^6.9.0",
     "react-scripts": "5.0.1",
     "react-select": "^5.7.2",
+    "react-share": "^4.4.1",
     "react-share-social": "^0.1.55",
     "redux-persist": "^6.0.0",
     "redux-thunk": "^2.4.2",
