+++ src/appComponents/DealModule/LatestDeals/LatestDeals.css
