+++ src/routes/Routes.js
@@ -8,14 +8,14 @@ import Login from "../authComponents/Login/Login";
 import LoginForm from "../authComponents/Login/LoginForm/LoginForm";
 import PasswordSent from "../authComponents/PasswordSent/PasswordSent";
 import SignupForm from "../authComponents/SignupForm/SignupForm";
-import LatestDeals from "../appComponents/LatestDeals/LatestDeals";
+import LatestDeals from "../appComponents/DealModule/LatestDeals/LatestDeals";
 import Profile from "../appComponents/Profile/Profile";
 import { useSelector } from "react-redux";
 import ViewCompanyProfile from "../appComponents/ViewCompanyProfile/ViewCompanyProfile";
-
 import ClassiFieds from "../appComponents/ClassiFieds/ClassiFieds";
 import PageNotFound from "../appComponents/PageNotFound/PageNotFound";
 import JobTypes from "../appComponents/ClassiFieds/JobTypes";
+
 
 
 //-------Create a component for manage routing--------
@@ -33,10 +33,10 @@ function Routers() {
                 <Route exact path="/forgot-password" element={<ForgotPassword />} />
                 <Route exact path="/password-sent" element={<PasswordSent />} />
                 <Route exact path="/email-varify" element={<EmailVarify />} />
-                <Route exact path="/deals/latest-deals" element={<LatestDeals />} />
+                <Route exact path="/deals/latest-deals" element={<LatestDeals/>} />
                 {/* <Route exact path="/deals/products" element={<Profile />} /> */}
                 <Route exact path="/view-profile" element={Object.keys(currentUser).length !== 0 ? (<Profile />) : (<Navigate to="/login" />)} />
-                <Route exact path="/deals/latest-deals/company-profile" element={<ViewCompanyProfile />} />
+                <Route exact path="/deals/latest-deals/company-profile/:id" element={<ViewCompanyProfile />} />
                 <Route exact path="/classifieds" element={<ClassiFieds />} />
                 <Route exact path="/job-types" element={<JobTypes/>} />
